KAN-146 refactor: 공통 타입 및 상수 분리
