+++ src/app/components/editor/ai-assistant-interface/ManualModification.tsx
@@ -5,15 +5,13 @@ import { ChangeEvent, useRef, useState } from 'react'
 import { TOAST_MESSAGE } from 'constants/common/toastMessage'
 import { FaCheck } from 'react-icons/fa6'
 import { IoClose } from 'react-icons/io5'
-import { ActionOptionType } from 'types/common/editor'
+import { ActionOptionType, EvaluateStateType } from 'types/common/editor'
 import { ModalHandler } from 'types/common/modalRef'
 
 import FillButton from '@components/buttons/FillButton'
 import Modal from '@components/modal/Modal'
 import SelectMenu from '@components/select-menu/SelectMenu'
 import { useToast } from '@components/toast/ToastProvider'
-
-import { EvaluateStateType } from '@hooks/editor/useTextEditor'
 
 import styles from '../DefaultEditor.module.scss'
