+++ src/app/constants/workspace/value.ts
@@ -1,0 +1,6 @@
+export const INITIAL_EVALUATE_STATE = {
+  assistantId: null,
+  isGoodSelected: false,
+  isBadSelected: false,
+  isArchived: false,
+}
